software/requirements.txt
matplotlib
flask
pyserial
opencv-python
ultralytics
face_recognition
filterpy
scipy

